+++ Thunderstore/CHANGELOG.md
@@ -1,4 +1,7 @@
 > # Update Information (Latest listed first)
+> ### 1.0.9
+> - Move to using the game's UnifiedPopup system. This should fix the issue with the popup.
+> - Add some new translations for the new popup and a few that I left out.
 > ### 1.0.8
 > - Fix issue with menu popup. Mouse now appears correctly.
 > - Prevent the popup from appearing if the player is in combat. This automatically denies the popup resulting in a denied sleep request.